Some structure and include changes (non-functional)
