+++ cli/CMakeLists.txt
@@ -1,8 +1,6 @@
 # SPDX-License-Identifier: Apache-2.0
 
 cmake_minimum_required(VERSION 3.13.1)
-list(APPEND ZEPHYR_EXTRA_MODULES
-   ${CMAKE_CURRENT_SOURCE_DIR}/../secure_module)
 
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(cli)
