+++ cli/prj.conf
@@ -40,7 +40,8 @@ CONFIG_NORDIC_SECURITY_BACKEND=y
 #CONFIG_SHELL_BACKEND_SERIAL_RX_RING_BUFFER_SIZE=512
 #CONFIG_SHELL_BACKEND_SERIAL_TX_RING_BUFFER_SIZE=512
 
-CONFIG_SPM=y
-CONFIG_SPM_BUILD_STRATEGY_SKIP_BUILD=y
-CONFIG_ARM_FIRMWARE_USES_SECURE_ENTRY_FUNCS=y
-CONFIG_ARM_ENTRY_VENEERS_LIB_NAME="spm/libspmsecureentries.a"
+# MAY SHOULD BE DELETED
+#CONFIG_SPM=y
+#CONFIG_SPM_BUILD_STRATEGY_SKIP_BUILD=y
+#CONFIG_ARM_FIRMWARE_USES_SECURE_ENTRY_FUNCS=y
+#CONFIG_ARM_ENTRY_VENEERS_LIB_NAME="spm/libspmsecureentries.a"
