+++ lib/bls_hsm_ns.c
@@ -50,105 +50,7 @@ int PBKDF2(uint8_t* salt, uint8_t* password, int it_cnt, uint8_t* key)
 }
 #endif
 
-/*
-Converts input 'pk' to binary 'out'
-*/
-void pk_serialize(byte* out, blst_p1 pk){
-        blst_p1_compress(out, &pk);
-}
 
-/*
-Converts input 'sig' to binary 'out2'
-*/
-void sig_serialize(byte* out2, blst_p2 sig){
-        blst_p2_compress(out2, &sig);
-}
-
-/*
-Get 'hash' from binary string 'msg_bin' with length 'len'
-*/
-void get_point_from_msg(blst_p2* hash, uint8_t* msg_bin, int len){
-        char dst[] = "BLS_SIG_BLS12381G2_XMD:SHA-256_SSWU_RO_POP_"; //IETF BLS Signature V4
-        //Obtain the point from a message
-        blst_hash_to_g2(hash, msg_bin, len, dst, sizeof(dst)-1, NULL, 0);
-}
-
-/*
-Derive 'pk' from hexadecimal string 'pk_hex'. Errors are dumped to 'buff'
-On error returns 1
-*/
-int pk_parse(char* pk_hex, blst_p1_affine* pk, char* buff){
-        byte pk_bin[48];
-        int offset = parse_hex(pk_hex, 96);
-        int error = 0;
-
-        if(offset == BADLEN){
-            strcat(buff, "Incorrect public key length. It must be 96 characters long.\n");
-            error = BADPKLEN;
-        }else if(offset == BADFORMAT){
-                strcat(buff, "Public key contains incorrect characters.\n");
-                error = offset;
-        }else{
-            if(hex2bin(pk_hex + offset, 96, pk_bin, 48) == 0) {
-                strcat(buff, "Failed converting public key to binary array\n");
-                error = HEX2BINERR;
-            }else{
-                blst_p1_uncompress(pk, pk_bin);
-            }
-        }
-        
-        return error;
-}
-
-/*
-Converts hexadecimal string 'msg' with length 'len' to binary string 'msg_bin'. Errors are dumped to 'buff'.
-On error returns 1
-*/
-int msg_parse(char* msg, uint8_t* msg_bin, int len, char* buff){
-
-        int offset = parse_hex(msg, len);
-        int error = 0;
-        if(offset == BADFORMAT){
-            strcat(buff, "Message contains incorrect characters.\n");
-            error = offset;
-        }else if(offset == BADLEN){
-            strcat(buff, "Message contains incorrect characters.\n");
-            error = offset;
-        }else{
-            if(hex2bin(msg + offset, len, msg_bin, len/2 + len%2) == 0) {
-                strcat(buff, "Incorrect message length.\n");
-                error = HEX2BINERR;
-            }
-        }
-
-        return error;
-}
-
-/*
-Derive 'sig' from hexadecimal string 'sig_hex'. Errors are dumped to 'buff'
-On error returns 1
-*/
-int sig_parse(char* sig_hex, blst_p2_affine* sig, char* buff){
-        byte sig_bin[96];
-        int offset = parse_hex(sig_hex, 192);
-        int error = 0;
-
-        if(offset == BADLEN){
-            strcat(buff, "Incorrect signature length. It must be 192 characters long.\n");
-            error = BADSIGLEN;
-        }else if(offset == BADFORMAT){
-            strcat(buff, "Signature contains incorrect characters.\n");
-            error = offset;
-        }else{
-            if(hex2bin(sig_hex + offset, 192, sig_bin, 96) == 0) {
-                strcat(buff, "Failed converting signature to binary array\n");
-                error = HEX2BINERR;
-            }else{
-                blst_p2_uncompress(sig, sig_bin);
-            }
-        }
-        return error;
-}
 
 /*
 Generates random key. Response is dumped to 'buff'
@@ -156,8 +58,12 @@ Generates random key. Response is dumped to 'buff'
 
 
 int keygen(char* data, char* buff){
+    #ifdef TFM
+    int keystore_size = tfm_get_keystore_size();
+    #else
     int keystore_size = get_keystore_size();
-
+    #endif
+    
     if(keystore_size < 10){
         // key_info is an optional parameter.  This parameter MAY be used to derive
         // multiple independent keys from the same IKM.  By default, key_info is the empty string.
@@ -166,8 +72,10 @@ int keygen(char* data, char* buff){
             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00};
+        int pk_index;
+        char pk[96];
 
-       //check if data string is empty
+        // Check if data string is empty
         if(strlen(data) != 0){
             if(strlen(data) <= strlen(info)){
                     strcpy(info, data);
@@ -176,24 +84,19 @@ int keygen(char* data, char* buff){
             }
         }
 
-        ikm_sk(info);
-    
-        //The secret key allow us to generate the associated public key
-        blst_p1 pk;
-        byte out[48];
-        char public_key_hex[96];
-        sk_to_pk(&pk);
-        pk_serialize(out, pk);
-        strcat(buff, "Public key: \n");
-        if(bin2hex(out, sizeof(out), public_key_hex, sizeof(public_key_hex)) == 0) {
-          strcat(buff, "Failed converting binary key to string\n");
-        return BIN2HEXERR;
+        // Generate sk and pk
+        pk_index = secure_keygen(info);
+        if(pk_index == -KEYSLIMIT){
+            strcat(buff, "Can't generate more keys. Limit reached.\n");
+            return pk_index;
+        }else if(pk_index == -BIN2HEXERR){
+            strcat(buff, "Error when converting public key from binary to hexadecimal.\n");
+            return pk_index;
         }
-        else
-        {
-        store_pk(public_key_hex);
-        print_pk(public_key_hex, buff);
+        if(get_key(pk_index, pk) != 0){
+            strcat(buff, "get_key: error\n");
         }
+        print_pk(pk, buff);
         
     }else{
         strcat(buff, "Can't generate more keys. Limit reached.\n");
@@ -203,7 +106,7 @@ int keygen(char* data, char* buff){
 
 /*
 Signs message with given public key. Public key must be stored. Response is dumped to 'buff'
-*/
+*//*
 int signature(char* pk, char* msg, char* buff){
     //Message examples
     //char * msg_hex = "5656565656565656565656565656565656565656565656565656565656565656";
@@ -245,7 +148,7 @@ return OK;
 
 /*
 Verifies signature of given message and public key
-*/
+*//*
 
 int verify(char* pk, char* msg, char* sig, char* buff){
     char dst[] = "BLS_SIG_BLS12381G2_XMD:SHA-256_SSWU_RO_POP_"; //IETF BLS Signature V4
@@ -267,7 +170,7 @@ int verify(char* pk, char* msg, char* sig, char* buff){
 }
 /*
 Get array of stored public keys in buffer 'buff'
-*/
+*//*
 
 int print_keys_Json(char* buff){
     int keystore_size = get_keystore_size();
@@ -292,7 +195,7 @@ int print_keys_Json(char* buff){
 
 /*
 Delete all stored public and secret keys. Response is dumped to 'buff'
-*/
+*//*
 
 
 void resetc(char* buff){
@@ -306,7 +209,7 @@ Import given secret key. Derived public key and errors are dumped to 'buff'
 
 /*
 Import given secret key. Derived public key and errors are dumped to 'buff'
-*/
+*//*
 int import(char* sk, char* buff){
     if(get_keystore_size() < 10){
         int offset = parse_hex(sk, 64);
@@ -353,7 +256,7 @@ int import(char* sk, char* buff){
 
 /*
     Returns 0 on succes and error number on error
-*/
+*//*
 int get_decryption_key_scrypt(char* password, int dklen, int n,  int r, int p, char* salt_hex, unsigned char* decryption_key){
     #ifndef NRF
     int error;
@@ -376,7 +279,7 @@ int get_decryption_key_scrypt(char* password, int dklen, int n,  int r, int p, c
 
 /*
     Returns 0 on succes and error number on error
-*/
+*//*
 int verify_password(char* checksum_message_hex, char* cipher_message_hex, unsigned char* decription_key){
     int error;
   
@@ -420,7 +323,7 @@ int verify_password(char* checksum_message_hex, char* cipher_message_hex, unsign
 
 /*
     Returns 0 on succes and error number on error
-*/
+*//*
 int get_private_key(char* cipher_message, char* iv_str, unsigned char* decription_key, char* private_key){
     int error;
 
@@ -455,6 +358,6 @@ int get_private_key(char* cipher_message, char* iv_str, unsigned char* decriptio
     import(private_key_str, buff);
 
     return 0;
-}
+}*/
