--- secure_module/zephyr/module.yml
@@ -1,3 +1,0 @@
-build:
-   cmake: zephyr
-#  kconfig: zephyr/Kconfig
