--- secure_module/zephyr/spm/boards/nrf5340dk_nrf5340_cpuapp.conf
@@ -1,1 +1,0 @@
-CONFIG_NRFX_UARTE2=y
