--- secure_module/zephyr/spm/boards/nrf5340dk_nrf5340_cpuapp.overlay
@@ -1,11 +1,0 @@
-/* SPDX-License-Identifier: LicenseRef-Nordic-5-Clause */
-
-&uart0 {
-	status = "okay";
-
-	spm_uart: nrf-spm-uart {
-		compatible = "nordic,nrf-spm-uart";
-		status = "okay";
-		label = "SPM_UART";
-	};
-};
