--- secure_module/zephyr/spm/boards/nrf9160dk_nrf9160.conf
@@ -1,1 +1,0 @@
-CONFIG_NRFX_UARTE2=y
