--- secure_module/zephyr/spm/nrf5340dk_nrf5340_cpuapp.overlay
@@ -1,12 +1,0 @@
-/* Needed to get NRF_PWMn defined. */
-&pwm1 {
-	status = "okay";
-};
-
-&pwm2 {
-	status = "okay";
-};
-
-&pwm3 {
-	status = "okay";
-};
