--- secure_module/zephyr/spm/nrf9160dk_nrf9160.overlay
@@ -1,22 +1,0 @@
-/* Needed to get NRF_PWMn defined. */
-&pwm1 {
-	status = "okay";
-};
-
-&pwm2 {
-	status = "okay";
-};
-
-&pwm3 {
-	status = "okay";
-};
-
-&uart0 {
-	status = "okay";
-
-	spm_uart: nrf-spm-uart {
-		compatible = "nordic,nrf-spm-uart";
-		status = "okay";
-		label = "SPM_UART";
-	};
-};
