--- secure_module/zephyr/spm/pm.yml
@@ -1,12 +1,0 @@
-#include <autoconf.h>
-
-spm:
-  size: 0x20000
-  placement:
-    before: app
-  inside: mcuboot_primary_app
-
-spm_sram:
-  size: 0x10000
-  placement: {after: start}
-  inside: sram_secure
