--- secure_module/zephyr/spm/prj.conf
@@ -1,26 +1,0 @@
-#
-# Copyright (c) 2019 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-CONFIG_NCS_SAMPLES_DEFAULTS=y
-
-CONFIG_IS_SPM=y
-CONFIG_FW_INFO=y
-CONFIG_GPIO=n
-CONFIG_MAIN_STACK_SIZE=25600
-
-# Images that set CONFIG_BOOTLOADER_MCUBOOT get this value set by default.
-# The SPM image will not have CONFIG_BOOTLOADER_MCUBOOT set by default when
-# being built by a parent image. Hence we set it here to ensure that SPM
-# cleans up the core during boot.
-CONFIG_INIT_ARCH_HW_AT_BOOT=y
-
-# Avoid using feature that allocates PPI channels
-CONFIG_UART_0_ENHANCED_POLL_OUT=n
-CONFIG_UART_1_ENHANCED_POLL_OUT=n
-
-CONFIG_ARM_FIRMWARE_HAS_SECURE_ENTRY_FUNCS=y
-CONFIG_SPM_SERVICE_RNG=y
-CONFIG_NORDIC_SECURITY_BACKEND=y
-CONFIG_NRF_OBERON=y
