--- secure_module/zephyr/spm/prj_minimal.conf
@@ -1,39 +1,0 @@
-#
-# Copyright (c) 2021 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-# Configurations needed for SPM sample to operate correctly.
-CONFIG_IS_SPM=y
-CONFIG_FW_INFO=y
-CONFIG_MAIN_STACK_SIZE=4096
-CONFIG_INIT_ARCH_HW_AT_BOOT=y
-
-# SPM specific optimizations
-CONFIG_SPM_MIN_PARTITION_SIZE=y
-CONFIG_SPM_BOOT_SILENTLY=y
-
-# General options to reduce code size
-CONFIG_ARM_MPU=n
-CONFIG_ASSERT=n
-CONFIG_BOOT_BANNER=n
-CONFIG_CONSOLE=n
-CONFIG_ERRNO=n
-CONFIG_GEN_IRQ_VECTOR_TABLE=n
-CONFIG_GPIO=n
-CONFIG_KERNEL_MEM_POOL=n
-CONFIG_LOG=n
-CONFIG_MULTITHREADING=n
-CONFIG_NCS_SAMPLES_DEFAULTS=n
-CONFIG_NRF_ENABLE_ICACHE=n
-CONFIG_PRINTK=n
-CONFIG_RESET_ON_FATAL_ERROR=n
-CONFIG_SECURE_BOOT_DEBUG=n
-CONFIG_SERIAL=n
-CONFIG_SYS_CLOCK_EXISTS=n
-CONFIG_SYS_POWER_MANAGEMENT=n
-CONFIG_TICKLESS_KERNEL=n
-CONFIG_TIMEOUT_64BIT=n
-CONFIG_UART_CONSOLE=n   # Set explicitly to 'n' to avoid Kconfig warning
-CONFIG_USE_SEGGER_RTT=n
