--- secure_module/zephyr/spm/sample.yaml
@@ -1,23 +1,0 @@
-sample:
-  name: spm Sample
-tests:
-  samples.spm:
-    build_only: true
-    build_on_all: true
-    platform_allow: nrf5340dk_nrf5340_cpuapp nrf9160dk_nrf9160
-    integration_platforms:
-      - nrf5340dk_nrf5340_cpuapp
-      - nrf9160dk_nrf9160
-    tags: ci_build
-
-  samples.spm_shared_uart:
-    build_only: true
-    build_on_all: true
-    platform_allow: nrf5340dk_nrf5340_cpuapp nrf9160dk_nrf9160
-    integration_platforms:
-      - nrf5340dk_nrf5340_cpuapp
-      - nrf9160dk_nrf9160
-    tags: ci_build
-    extra_configs:
-      - CONFIG_SPM_SHARE_CONSOLE_UART=y
-      - CONFIG_UART_CONSOLE_ON_DEV_NAME="SPM_UART"
