--- secure_module/zephyr/spm/spm_enable_cpunet.conf
@@ -1,1 +1,0 @@
-CONFIG_BOARD_ENABLE_CPUNET=y
