--- secure_module/zephyr/spm/spm_ns_handler_from_spm_fault.conf
@@ -1,17 +1,0 @@
-#
-# Copyright (c) 2021 Nordic Semiconductor ASA
-#
-# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
-#
-
-# Set of options applied to SPM when CONFIG_SPM_SERVICE_NS_HANDLER_FROM_SPM_FAULT
-# is enabled in non-secure application.
-
-# When enabled, spm can print to UARTE0 even after switching to non-secure, e.g.
-# from secure fault handler.
-CONFIG_SPM_SHARE_CONSOLE_UART=y
-CONFIG_UART_CONSOLE_ON_DEV_NAME="SPM_UART"
-
-# When enabled, spm, in fault handler, will call non-secure fatal error handler.
-CONFIG_SPM_SERVICE_NS_HANDLER_FROM_SPM_FAULT=y
-CONFIG_SPM_SECURE_SERVICES=y
